Using auth layers

+++ app/src/main/java/com/github/dannful/uopoomsae/domain/model/ScoreData.kt
@@ -4,8 +4,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ScoreData(
-    val tableId: Int,
-    val judgeId: Int,
+    val tableId: Short,
+    val judgeId: Short,
     val presentationScore: Float,
     val accuracyScore: Float
 )
